refactor(SRP): Move loan application related logic into its own service
